Algorithms.BF.QuantumIf

Description

This module introduces a simple way of defining "boolean statements" acting over qubits. See Algorithms.BF.Main for an overview of the boolean formula algorithm.

Quantum "if then else" statements

This section introduces a simple way of defining "boolean statements" acting over qubits, that can then be used as the control in a quantum "If Then Else" statement. The idea is that an ancilla is initialized that is in the state represented by the boolean statement, and is then used to control the two branches of the "If Then Else", before being uncomputed. The boolean statements can contain "and", "or", and "not".

data Boolean a

We can use (Boolean Qubit) to build up "boolean statements" over qubits and use the "boolean statement" in an if_then_elseQ construct.

 Example (for qubits a, b, c, d):
 (a and b) or !(c and !d) is written as:
 Or (And (A a) (A b)) (Not (And (A c) (Not (A d)))) 

Constructors

A a
A q means if q == True.

Not (Boolean a)
Not b means the negation of the boolean statement b.

And (Boolean a) (Boolean a)
And a b means the and of the boolean statements a and b.

Or (Boolean a) (Boolean a)
Or a b means the or of the boolean statements a and b

data BooleanAnd a

Allow And and Or to be used as infix operators, with the same precedences.

Internally, a "boolean statement" is converted into a statement that doesn't use or (e.g., using De Morgan's laws).

Constructors

AA a
AA q means if q == True.

NotA (BooleanAnd a)
NotA b means the negation of the boolean statement b.

AndA (BooleanAnd a) (BooleanAnd a)
AndA a b means the and of the boolean statements a and b.

booleanToAnd' :: Boolean a -> BooleanAnd a

Convert any boolean formula to a formula using just and and not. This conversion function uses De Morgan's law, i.e.,

 A or B = !( !A and !B ),

but does not remove double negations. For a version that also removes double negations, see booleanToAnd.

stripDoubleNot :: BooleanAnd a -> BooleanAnd a

Strip any redundant double negations, i.e., in this context !!A = A.

booleanToAnd :: Boolean a -> BooleanAnd a

Convert any boolean formula to a formula using just and and not, removing double negations.

booleanAnd' :: BooleanAnd Qubit -> Qubit -> Circ ()

Create a circuit from the "boolean statement".

booleanAnd :: BooleanAnd Qubit -> Circ Qubit

Create a circuit from the "boolean statement", passing in an ancilla.

if_then_elseQinv :: Boolean Qubit -> Circ () -> Circ () -> Boolean Qubit -> Circ ()

The definition of a quantum if_then_else structure uses a "boolean statement" to create a single ancilla in the state defined by the boolean statement, and uses this as a control for the two branches of the if statement. The ancilla then needs to be uncomputed, this is achieved using the other given "boolean statement", i.e., a new boolean statement that would produce the state of the control ancilla, from the output state of the two branches.This allows the branches to update the state of qubits used in the original "boolean statement" as long as it is done in a (reversible) known-manner. This is useful for the WALK algorithm, where TOPARENT and TOCHILD are controlled by the state of the direction register, but also change the state of the direction register.

if_then_elseQ :: Boolean Qubit -> Circ () -> Circ () -> Circ ()

Like if_then_elseQinv, but where the original "boolean statement" still holds after the branches have taken place.