Algorithms.QLS.CircLiftingImport

Description

This module contains definitions to work with Template Haskell. All the definitions in this module are used by Template Haskell in Algorithms.QLS.TemplateOracle and Algorithms.QLS.RealFunc.

Utility function

grepn :: Eq a => [a] -> [a] -> Int

grepn regexp list: Counts how many times regexp is a sublist of list.

Lifting of ordering operators.

template_symb_slash_symb_equal_ :: (Typeable qa, QOrd qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ Qubit))

Template version of /=.

template_symb_oangle_ :: (Typeable qa, QOrd qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ Qubit))

Template version of <.

template_symb_oangle_symb_equal_ :: (Typeable qa, QOrd qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ Qubit))

Template version of <=.

template_symb_cangle_ :: (Typeable qa, QOrd qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ Qubit))

Template version of >.

template_symb_cangle_symb_equal_ :: (Typeable qa, QOrd qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ Qubit))

Template version of >=.

Lifting of arithmetic operators

template_symb_minus_ :: (Typeable qa, QNum qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ qa))

Template version of -.

template_symb_plus_ :: (Typeable qa, QNum qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ qa))

Template version of +.

template_symb_star_ :: (Typeable qa, QNum qa, Show (BType qa)) => Circ (qa -> Circ (qa -> Circ qa))

Template version of *.

template_negate :: (Typeable qa, QNum qa, Show (BType qa)) => Circ (qa -> Circ qa)

Template version of negate.

template_abs :: (Typeable qa, QNum qa, Show (BType qa)) => Circ (qa -> Circ qa)

Template version of abs.

template_mod :: Circ (QSignedInt -> Circ (QSignedInt -> Circ QSignedInt))

Template version of mod

Operations on QDouble

template_symb_slash_ :: Circ (QDouble -> Circ (QDouble -> Circ QDouble))

Template version of / on Fractional.

local_pi :: FDouble

The constant pi as an FDouble.

template_local_pi :: Circ QDouble

Template version of local_pi.

Relation between QDouble and QSignedInt.

template_floor :: Circ (QDouble -> Circ QSignedInt)

Template version of floor.

template_ceiling :: Circ (QDouble -> Circ QSignedInt)

Template version of ceil.

template_fromIntegral :: Circ (QSignedInt -> Circ QDouble)

Template version of fromIntegral.

Dealing with parameters.

template_rational :: Double -> Circ QDouble

Lift a real number to QDouble.

template_integer :: Int -> Circ QSignedInt

Lift an integer to QSignedInt.

getIntFromParam :: Int -> Int

Make a parameter Int as a regular Int that can be lifted.

template_getIntFromParam :: Circ (Int -> Circ QSignedInt)

Template version of getIntFromParam.

paramZero :: Int

Parameter integer of value '0'.

template_paramZero :: Circ Int

Template version of paramZero.

paramTen :: Int

Parameter integer of value '10'.

template_paramTen :: Circ Int

Template version of paramTen.

paramSucc :: Int -> Int

Successor function acting on parameter Int.

template_paramSucc :: Circ (Int -> Circ Int)

Template version of paramSucc.

paramPred :: Int -> Int

Predecessor function acting on parameter Int.

template_paramPred :: Circ (Int -> Circ Int)

Template version of paramPred.

paramMinus :: Int -> Int -> Int

Subtraction of parameter integers.

template_paramMinus :: Circ (Int -> Circ (Int -> Circ Int))

Template version of paramMinus.

Miscellaneous operations.

template_length :: Circ ([a] -> Circ QSignedInt)

Lifted version of length.

take_half :: [a] -> [a]

Return the first half of the input list.

template_take_half :: Circ ([a] -> Circ [a])

Lifted version of take_half.