Algorithms.QLS.QSignedIntAux

Description

Helper module for Algorithms.QLS.QSignedInt: some functions defined with Template Haskell.

Documentation

be_boollist_sub :: [Bool] -> [Bool] -> [Bool]

Subtraction on lists of booleans, understood as big-headian, unsigned integers.

be_boollist_add :: [Bool] -> [Bool] -> [Bool]

Addition on lists of booleans, understood as big-headian, unsigned integers.

be_boollist_less :: [Bool] -> [Bool] -> Bool

Strict ordering on lists of booleans, understood as big-headian unsigned integers. If the lists are not of equal length, the shorter list is treated as if its tail were padded with zeros.

be_signed_boollist_less :: (Bool, [Bool]) -> (Bool, [Bool]) -> Bool

Strict ordering on lists of booleans, understood as big-headian signed integers: the Bool in the pair stands for the sign: False is positive, True is negative.

boollist_is_zero :: [Bool] -> Bool

Test whether all elements of a list are False.

be_signed_boollist_add :: (Bool, [Bool]) -> (Bool, [Bool]) -> (Bool, [Bool])

Addition on signed integers, encoded as big-headian lists of booleans.