Algorithms.QLS.Utils

Description

This module defines some auxiliary machinery required for the QLS algorithm.

Hard-coded default sizes for quantum numbers

fixed_int_register_length :: Int

Default size of a register QSignedInt (not counting the sign).

before_radix_length :: Int

Default size for the xxx part of the QDouble xxx.yyy.

after_radix_length :: Int

Default size for the yyy part of the QDouble xxx.yyy.

Miscellaneous utilities

ncompose :: Int -> (a -> a) -> a -> a

Compose a function with itself n times.

listpair_fmap :: (a -> b) -> [(a, a)] -> [(b, b)]

Specialized map for lists of pairs.