Libraries.Stabilizers.Clifford

Description

This module contains an implementation of a quantum simulator that uses the stabilizer states of the Clifford group (i.e. the Pauli group), to provide efficient simulation of quantum circuits constructed from elements of the Clifford group. The module provides an implementation of the Clifford group operators {x,y,z,h,s,controlled-x} which form a generating set for the Clifford group.

Documentation

type Qubit = Int

A qubit is defined as an integer reference.

data Tableau

The state of the system is a representation of a stabilizer tableau.

Constructors

ST Qubit !(Map Qubit Sign) !(Map (Qubit, Qubit) Pauli) !(Map Qubit Sign) !(Map (Qubit, Qubit) Pauli)

next_qubit :: Tableau -> Qubit

Accessor function for the next_qubit field of a Tableau

sign :: Tableau -> Map Qubit Sign

Accessor function for the sign field of a Tableau

tableau :: Tableau -> Map (Qubit, Qubit) Pauli

Accessor function for the tableau field of a Tableau

de_sign :: Tableau -> Map Qubit Sign

Accessor function for the de_sign field of a Tableau

de_tableau :: Tableau -> Map (Qubit, Qubit) Pauli

Accessor function for the de_tableau field of a Tableau

lookup :: (Ord k, Show k, Show v) => k -> Map k v -> v

A local Map lookup function that throws an error if the key doesn't exist.

empty_tableau :: Tableau

An initial (empty) tableau.

add_qubit :: Bool -> Tableau -> Tableau

A new qubit in the state |0〉 or |1〉 can be added to a tableau.

type Unitary = Pauli -> (Sign, Pauli)

A (Clifford) unitary can be defined as a function acting on Pauli operators.

data MinPauli

The minimal definition of a unitary requires the actions on X and Z.

Constructors

Xmin

Zmin

type MinUnitary = MinPauli -> (Sign, Pauli)

The minimal definition of a unitary requires the actions on X and Z.

from_minimal :: MinUnitary -> Unitary

The definition of a Unitary can be constructed from a MinimalUnitary.

from_matrix :: (Floating r, Eq r, Show r) => Matrix1 (Cplx r) -> Unitary

It is possible to construct a Unitary from a 2×2-matrix.

apply_unitary :: Unitary -> Qubit -> Tableau -> Tableau

A unitary can be applied to a qubit in a given tableau. By folding through each row

apply_unitary_row :: Unitary -> Qubit -> Tableau -> Qubit -> Tableau

Apply the unitary to the given column, in the given row.

type Unitary2 = (Pauli, Pauli) -> (Sign, Pauli, Pauli)

A two-qubit (Clifford) unitary can be defined as a function acting on a pair of Pauli operators.

data MinPauli2

The minimal definition of a two-qubit unitary requires the actions on IX, XI, IZ, and ZI.

Constructors

IX

XI

IZ

ZI

type MinUnitary2 = MinPauli2 -> (Sign, Pauli, Pauli)

The minimal definition of a two-qubit unitary requires the actions on IX, XI, IZ, and ZI.

from_minimal2 :: MinUnitary2 -> Unitary2

The definition of a Unitary2 can be constructed from a MinimalUnitary2.

from_matrix2 :: (Floating r, Eq r, Show r) => Matrix2 (Cplx r) -> Unitary2

It is possible to construct a Unitary2 from a 4×4-matrix.

from_matrix_controlled :: (Floating r, Show r, Eq r) => Matrix1 (Cplx r) -> Unitary2

It is possible to construct a Unitary2 from controlling a 2×2-matrix.

apply_unitary2 :: Unitary2 -> (Qubit, Qubit) -> Tableau -> Tableau

A two-qubit unitary can be applied to a pair of qubits in a given tableau.

measure :: Qubit -> Tableau -> IO (Bool, Tableau)

A measurement, in the computational basis, can be made of a qubit in the Tableau, returning the measurement result, and the resulting Tableau.

reduce :: Qubit -> Tableau -> Tableau

This function reduces a tableau so that it contains either plus or minus Zq. Note that it is only called in the case where Zq is generated by the tableau (i.e., during measurement).

multiply :: Qubit -> Qubit -> Tableau -> Tableau

Multiply the stabilizers for the two given rows, in the given tableau, and update the first row with the result of the multiplication.

x :: Unitary

The Pauli X operator is a Clifford group unitary.

x_min :: MinUnitary

We can (equivalently) define Pauli-X as a MinUnitary.

x' :: Unitary

We can (equivalently) construct Pauli-X from a MinUnitary.

x'' :: Unitary

We can (equivalently) construct Pauli-X from a matrix.

y :: Unitary

The Pauli Y-operator is a Clifford group unitary.

y_min :: MinUnitary

We can (equivalently) define Pauli-Y as a MinUnitary.

y' :: Unitary

We can (equivalently) construct Pauli-Y from a MinUnitary.

y'' :: Unitary

We can (equivalently) construct Pauli-Y from a matrix.

z :: Unitary

The Pauli Z-operator is a Clifford group unitary.

z_min :: MinUnitary

We can (equivalently) define Pauli-Z as a MinUnitary.

z' :: Unitary

We can (equivalently) construct Pauli-Z from a MinUnitary.

z'' :: Unitary

We can (equivalently) construct Pauli-Z from a matrix.

h :: Unitary

The Hadamard-gate is a Clifford group unitary.

h_min :: MinUnitary

We can (equivalently) define Hadamard as a MinUnitary.

h' :: Unitary

We can (equivalently) construct Hadamard from a MinUnitary.

h'' :: Unitary

We can (equivalently) construct Hadamard from a matrix. Although rounding errors break this!!!

s :: Unitary

The phase-gate is a Clifford group unitary.

s_min :: MinUnitary

We can (equivalently) define phase gate as a MinUnitary.

s' :: Unitary

We can (equivalently) construct phase gate from a MinUnitary.

s'' :: Unitary

We can (equivalently) construct phase gate from a matrix.

e :: Unitary

The phase-gate is a Clifford group unitary.

e_min :: MinUnitary

We can (equivalently) define phase gate as a MinUnitary.

e' :: Unitary

We can (equivalently) construct phase gate from a MinUnitary.

e'' :: Unitary

We can (equivalently) construct phase gate from a matrix.

cnot :: Unitary2

The controlled-not is a Clifford group 2-qubit unitary.

cnot_min :: MinUnitary2

We can (equivalently) define CNot as a MinUnitary2.

cnot' :: Unitary2

We can (equivalently) construct CNot from a MinUnitary2.

cnot'' :: Unitary2

We can (equivalently) construct CNot from a matrix.

cz :: Unitary2

The controlled-Z is a Clifford group 2-qubit unitary.

cz_min :: MinUnitary2

We can (equivalently) define controlled-Z as a MinUnitary2.

cz' :: Unitary2

We can (equivalently) construct controlled-Z from a MinUnitary2.

cz'' :: Unitary2

We can (equivalently) construct controlled-Z from a matrix.

type CliffordCirc a = StateT Tableau IO a

A Clifford group circuit is implicitly simulated using a state monad over a Tableau.

init_qubit :: Bool -> CliffordCirc Qubit

Initialize a new qubit.

init_qubits :: [Bool] -> CliffordCirc [Qubit]

Initialize multiple qubits.

gate_X :: Qubit -> CliffordCirc ()

Apply a Pauli-X gate to the given qubit.

gate_Y :: Qubit -> CliffordCirc ()

Apply a Pauli-Y gate to the given qubit.

gate_Z :: Qubit -> CliffordCirc ()

Apply a Pauli-Z gate to the given qubit.

gate_H :: Qubit -> CliffordCirc ()

Apply a Hadamard gate to the given qubit.

gate_S :: Qubit -> CliffordCirc ()

Apply a phase gate to the given qubit.

gate_Unitary :: Unitary -> Qubit -> CliffordCirc ()

Apply a given Unitary to the given qubit.

controlled_X :: Qubit -> Qubit -> CliffordCirc ()

Apply a controlled-X gate to the given qubits.

controlled_Z :: Qubit -> Qubit -> CliffordCirc ()

Apply a controlled-Z gate to the given qubits.

gate_Unitary2 :: Unitary2 -> Qubit -> Qubit -> CliffordCirc ()

Apply a given Unitary2 to the given qubits

measure_qubit :: Qubit -> CliffordCirc Bool

Measure the given qubit in the computational basis.

measure_qubits :: [Qubit] -> CliffordCirc [Bool]

Measure the given list of qubits.

show_tableau :: CliffordCirc ()

For testing purposes, we can show the tableau during a simulation.

eval :: CliffordCirc a -> IO Tableau

Return the evaluated Tableau for the given circuit.

sim :: CliffordCirc a -> IO a

Return the result of simulating the given circuit.

swap :: Qubit -> Qubit -> CliffordCirc ()

A swap gate can be defined in terms of three controlled-not gates.

controlled_Z' :: Qubit -> Qubit -> CliffordCirc ()

A controlled-Z gate can (equivalently) be defined in terms of Hadamard and controlled-X.

bell :: (Bool, Bool) -> CliffordCirc (Qubit, Qubit)

Each of the four Bell states can be generated, indexed by a pair of boolean values.

measure_bell00 :: CliffordCirc (Bool, Bool)

Create a Bell state, and measure it.

controlled_if :: Bool -> (Qubit -> CliffordCirc ()) -> Qubit -> CliffordCirc ()

A single-qubit operation can be controlled by a classical boolean value.

teleport :: Qubit -> CliffordCirc Qubit

A simple, single qubit, teleportation circuit.

test_teleport :: Bool -> CliffordCirc Bool

A wrapper around the teleportation circuit that initializes a qubit in the given boolean state, and measures the teleported qubit.

random_bool :: CliffordCirc Bool

Measure an equal superposition.