Libraries.Synthesis.LaTeX

Description

This module provides some functionality for pretty-printing certain types to LaTeX format.

Documentation

class ShowLaTeX a where

A type class for things that can be printed to LaTeX format.

Minimal complete definition: showlatex or showlatex_p.

Methods

showlatex :: a -> String

Print to LaTeX format.

showlatex_p :: Int -> a -> ShowS

Print to LaTeX format, with precedence. Analogous to showsPrec.

showlatex_denomexp :: (WholePart a b, ShowLaTeX b, DenomExp a) => a -> String

Generic showlatex-like method that factors out a common denominator exponent.