Libraries.Synthesis.Ring

Description

This module provides type classes for rings. It also provides several specific instances of rings, such as the ring ℤ₂ of integers modulo 2, the ring ℚ of rational numbers, the ring ℤ[½] of dyadic fractions, the ring ℤ[i] of Gaussian integers, the ring ℤ[√2] of quadratic integers with radix 2, and the ring ℤ[ω] of cyclotomic integers of degree 8.

Rings

class Num a => Ring a

A type class to denote rings. We make Ring a synonym of Haskell's Num type class, so that we can use the usual notation +, -, * for the ring operations. This is not a perfect fit, because Haskell's Num class also contains two non-ring operations abs and signum. By convention, for rings where these notions don't make sense (or are inconvenient to define), we set abs x = x and signum x = 1.

Rings with particular elements

We define several classes of rings with special elements.

Rings with ½

class Ring a => HalfRing a where

A type class for rings that contain ½.

Methods

half :: a

The value ½.

Rings with √2

class Ring a => RootTwoRing a where

A type class for rings that contain √2.

Methods

roottwo :: a

The square root of 2.

Rings with 1/√2

class (HalfRing a, RootTwoRing a) => RootHalfRing a where

A type class for rings that contain 1/√2.

Methods

roothalf :: a

The square root of ½.

Rings with i

class Ring a => ComplexRing a where

A type class for rings that contain a square root of -1.

Methods

i :: a

The complex unit.

Rings with ω

class Ring a => OmegaRing a where

A type class for rings that contain a square root of i, or equivalently, a fourth root of −1.

Methods

omega :: a

The square root of i.

Rings with particular automorphisms

Rings with complex conjugation

class Adjoint a where

A type class for rings with complex conjugation, i.e., an automorphism mapping i to −i.

When instances of this type class are vectors or matrices, the conjugation also exchanges the roles of rows and columns (in other words, it is the adjoint).

For rings that are not complex, the conjugation can be defined to be the identity function.

Methods

adj :: a -> a

Compute the adjoint (complex conjugate transpose).

Rings with √2-conjugation

class Adjoint2 a where

A type class for rings with a √2-conjugation, i.e., an automorphism mapping √2 to −√2.

When instances of this type class are vectors or matrices, the √2-conjugation does not exchange the roles of rows and columns.

For rings that have no √2, the conjugation can be defined to be the identity function.

Methods

adj2 :: a -> a

Compute the adjoint, mapping a + b√2 to a −b√2.

Normed rings

class Ring r => NormedRing r where

A (number-theoretic) norm on a ring R is a function N : R → ℤ such that N(rs) = N(r)N(s), for all r, s ∈ R. The norm also satisfies N(r) = 0 iff r = 0, and N(r) = ±1 iff r is a unit of the ring.

Methods

norm :: r -> Integer

Floor and ceiling

class Ring r => Floor r where

The floor and ceiling functions provided by the standard Haskell libraries are predicated on many unnecessary assumptions. This type class provides an alternative.

Minimal complete definition: floor_of or ceiling_of.

Methods

floor_of :: r -> Integer

Compute the floor of x, i.e., the greatest integer n such that n ≤ x.

ceiling_of :: r -> Integer

Compute the ceiling of x, i.e., the least integer n such that x ≤ n.

Particular rings

The ring ℤ₂ of integers modulo 2

data Z2

The ring ℤ₂ of integers modulo 2.

Constructors

Even

Odd

The ring D of dyadic rationals

data Dyadic

A dyadic rational is a rational number whose denominator is a power of 2. We denote the dyadic rationals by D = ℤ[½].

We internally represent a dyadic rational a/2n as a pair (a,n). Note that this representation is not unique. When it is necessary to choose a canonical representative, we choose the least possible n≥0.

Constructors

Dyadic !Integer !Integer

decompose_dyadic :: Dyadic -> (Integer, Integer)

Given a dyadic rational r, return (a,n) such that r = a/2n, where n≥0 is chosen as small as possible.

integer_of_dyadic :: Dyadic -> Integer -> Integer

Given a dyadic rational r and an integer k≥0, such that a = r2k is an integer, return a. If a is not an integer, the behavior is undefined.

fromDyadic :: HalfRing a => Dyadic -> a

The unique ring homomorphism from ℤ[½] to any ring containing ½. This exists because ℤ[½] is the free such ring.

The ring ℚ of rational numbers

newtype Rationals

We define our own variant of the rational numbers, which is an identical copy of the type Rational from the standard Haskell library, except that it has a more sensible Show instance.

Constructors

ToRationals

Fields

unRationals :: Rational

showsPrec_rational :: (Show a, Integral a) => Int -> Ratio a -> ShowS

An auxiliary function for printing rational numbers, using correct precedences, and omitting denominators of 1.

fromRationals :: Fractional a => Rationals -> a

Conversion from Rationals to any Fractional type.

The ring R[√2]

data RootTwo a

The ring R[√2], where R is any ring. The value RootTwo a b represents a + b √2.

Constructors

RootTwo !a !a

The ring ℤ[√2]

type DInteger = RootTwo Integer

The ring ℤ[√2].

fromDInteger :: RootTwoRing a => DInteger -> a

The unique ring homomorphism from ℤ[√2] to any ring containing √2. This exists because ℤ[√2] is the free such ring.

The ring ℤ[1/√2]

type DReal = RootTwo Dyadic

The ring ℤ[1/√2]. We think of the elements of ℤ[1/√2] as (certain) exact real numbers. The "D" stands for "dyadic".

fromDReal :: RootHalfRing a => DReal -> a

The unique ring homomorphism from ℤ[1/√2] to any ring containing 1/√2. This exists because ℤ[1/√2] is the free such ring.

The field ℚ[√2]

type EReal = RootTwo Rationals

The field ℚ[√2]. We think of the elements of ℚ[√2] as (certain) exact real numbers, so the "E" stands for "exact".

fromEReal :: (RootTwoRing a, Fractional a) => EReal -> a

The unique ring homomorphism from ℚ[√2] to any ring containing the rational numbers and √2. This exists because ℚ[√2] is the free such ring.

The ring R[i]

data Cplx a

The ring R[i], where R is any ring. The reason we do not use the Complex a type from the standard Haskell libraries is that it assumes too much, for example, it assumes a is a member of the RealFloat class. Also, this allows us to define a more sensible Show instance.

Constructors

Cplx !a !a

The ring ℤ[i] of Gaussian integers

type ZComplex = Cplx Integer

The ring ℤ[i] of Gaussian integers.

fromZComplex :: ComplexRing a => ZComplex -> a

The unique ring homomorphism from ℤ[i] to any ring containing i. This exists because ℤ[i] is the free such ring.

The ring ℤ[1/√2, i]

type DComplex = Cplx DReal

The ring ℤ[1/√2, i]. We think of the elements as (certain) exact complex numbers. The "D" stands for "dyadic".

fromDComplex :: (RootHalfRing a, ComplexRing a) => DComplex -> a

The unique ring homomorphism from ℤ[1/√2, i] to any ring containing 1/√2 and i. This exists because ℤ[1/√2, i] is the free such ring.

The ring ℚ[√2, i]

type EComplex = Cplx EReal

The field ℚ[√2, i]. We think of the elements as (certain) exact complex numbers, hence the "E" in the name.

fromEComplex :: (RootTwoRing a, ComplexRing a, Fractional a) => EComplex -> a

The unique ring homomorphism from ℚ[√2, i] to any ring containing the rational numbers, √2, and i. This exists because ℚ[√2, i] is the free such ring.

The ring ℂ of complex numbers

We provide two versions of the complex numbers using floating point arithmetic.

type CDouble = Cplx Double

Double precision complex floating point numbers.

type CFloat = Cplx Float

Single precision complex floating point numbers.

The ring R[ω]

data Omega a

The ring R[ω], where R is any ring, and ω is an 8th root of unity. The value Omega a b c d represents aω3+bω2+cω+d.

Constructors

Omega !a !a !a !a

omega_real :: Omega a -> a

An inverse to the embedding R ↦ R[ω]: return the "real rational" part.

The ring ℤ[ω]

type ZOmega = Omega Integer

The ring ℤ[ω] of cyclotomic integers of degree 8. Such rings were first studied by Kummer around 1840, and used in his proof of special cases of Fermat's Last Theorem. See also:

http://fermatslasttheorem.blogspot.com/2006/05/basic-properties-of-cyclotomic.html

http://fermatslasttheorem.blogspot.com/2006/02/cyclotomic-integers.html

Harold M. Edwards, "Fermat's Last Theorem: A Genetic Introduction to Algebraic Number Theory"

fromZOmega :: OmegaRing a => ZOmega -> a

The unique ring homomorphism from ℤ[ω] to any ring containing ω. This exists because ℤ[ω] is the free such ring.

dinteger_of_zomega :: ZOmega -> DInteger

Inverse of the embedding ℤ[√2] → ℤ[ω]. Note that ℤ[√2] = ℤ[ω] ∩ ℝ. This function takes an element of ℤ[ω] that is real, and converts it to an element of ℤ[√2]. It throws an error if the input is not real.

The ring D[ω]

type DOmega = Omega Dyadic

The ring D[ω]. Here D=ℤ[½] is the ring of dyadic fractions. In fact, D[ω] is isomorphic to the ring ℤ[1/√2, i], but they have different Show instances.

fromDOmega :: (RootHalfRing a, ComplexRing a) => DOmega -> a

The unique ring homomorphism from D[ω] to any ring containing 1/√2 and i. This exists because D[ω] is the free such ring.

The field ℚ[ω]

type QOmega = Omega Rationals

The field ℚ[ω] of cyclotomic rationals of degree 8.

Conversion to dyadic

class ToDyadic a b | a -> b where

A type class relating "rational" types to their dyadic counterparts.

Methods

maybe_dyadic :: a -> Maybe b

Convert a "rational" value to a "dyadic" value, if the denominator is a power of 2. Otherwise, return Nothing.

to_dyadic :: ToDyadic a b => a -> b

Convert a "rational" value to a "dyadic" value, if the denominator is a power of 2. Otherwise, throw an error.

Real part

class RealPart a b | a -> b where

A type class for rings that have a "real" component. A typical instance is a = DComplex with b = DReal.

Methods

real :: a -> b

Take the real part.

Rings of integers

class WholePart a b | a -> b where

A type class for rings that have a distinguished subring "of integers". A typical instance is a = DReal, which has b = DInteger as its ring of integers.

Methods

from_whole :: b -> a

The embedding of the ring of integers into the larger ring.

to_whole :: a -> b

The inverse of from_whole. Throws an error if the given element is not actually an integer in the ring.

Common denominators

class DenomExp a where

A type class for things from which a common power of 1/√2 (a least denominator exponent) can be factored out. Typical instances are DReal, DComplex, as well as tuples, lists, vectors, and matrices thereof.

Methods

denomexp :: a -> Integer

Calculate the least denominator exponent k of a. Returns the smallest k≥0 such that a = b/√2k for some integral b.

denomexp_factor :: a -> Integer -> a

Factor out a kth power of 1/√2 from a. In other words, calculate a√2k.

denomexp_decompose :: (WholePart a b, DenomExp a) => a -> (b, Integer)

Calculate and factor out the least denominator exponent k of a. Return (b,k), where a = b/(√2)k and k≥0.

showsPrec_DenomExp :: (WholePart a b, Show b, DenomExp a) => Int -> a -> ShowS

Generic show-like method that factors out a common denominator exponent.

Conversion to EComplex

EComplex is the largest one of our "exact" arithmetic types. We define a toEComplex family of functions for converting just about anything to an EComplex.

class ToEComplex a where

A type class for things that can be exactly converted to EComplex.

Methods

toEComplex :: a -> EComplex

Conversion to EComplex.

Parity

class Parity a where

A type class for things that have parity.

Methods

parity :: a -> Z2

Return the parity of something.

Auxiliary functions

lobit :: Integer -> (Integer, Integer)

If n≠0, return (a,k) such that a is odd and n = a⋅2k. If n=0, return (0,0).

log2 :: Integer -> Maybe Integer

If n is of the form 2k, return k. Otherwise, return Nothing.

intsqrt :: Integral n => n -> n

For n ≥ 0, return the floor of the square root of n. This is done using integer arithmetic, so there are no rounding errors.