Libraries.Tuple

Description

This module provides isomorphisms between n-tuples and repeated pairs. It is used to be able to write type classes for n-tuples more generically. Essentially we want to be able to write code for 17-tuples once and for all, rather than once for each type class we define. Ideally there would be a standard Haskell library for this.

Two type classes are provided: Tuple, and TupleOrUnary. Tuple is recommended for most uses.

Documentation

class TupleOrUnary t s | s -> t where

This type class relates types of the form t = (a,b,c,d) (“tupled form”) to types of the form s = (a,(b,(c,(d,())))) (“standard form”), and provides a way to convert between the two representations.

The tupled form can always be deduced from the standard form.

Methods

weak_tuple :: s -> t

For example, maps (a,(b,(c,(d,())))) to (a,b,c,d).

weak_untuple :: t -> s

For example, maps (a,b,c,d) to (a,(b,(c,(d,())))).

class TupleOrUnary t s => Tuple t s | s -> t, t -> s where

In almost all instances, the standard form can also be deduced from the tupled form; the only exception is the unary case. The Tuple class includes no new methods, adding just this functional dependency.

While the methods of Tuple are always copied from those of TupleOrUnary, they are renamed, so that use of these methods tells the type checker it can use the extra functional dependency.

Methods

tuple :: s -> t

untuple :: t -> s