Libraries.Typeable

Description

The standard Haskell module Data.Typeable only provides instances for tuples up to length 7. Since we need larger tuples, we provide the missing instances here.

Unfortunately, there is no easy way to do this portably; once the instances get added to Data.Typeable, we must remove them here.

Please note: type instances do not generate documentation, so there is nothing here to document. Please click on "Source" above to see the source code.