Quipper.Algorithms.BF.HexBoard

Description

This module provides the code for drawing Hex boards in graphical format. See Quipper.Algorithms.BF.Main for an overview of the boolean formula algorithm.

Documentation

output_start_board :: Format -> BooleanFormulaOracle -> IO ()

Output the starting HexBoard in the given format, for the given oracle.

output_HexBoards :: Format -> BooleanFormulaOracle -> [HexBoard] -> IO ()

Output multiple HexBoards in the given format, for the given oracle.