Quipper.Algorithms.CL.Test

Description

Test the Class Number algorithm, and its components, using classical computation

Sample data

Some fairly arbitrarily chosen sample elements of various types, for convenience in testing functions.

Matrices

sample_matrix :: CLMatrix Integer

A sample square matrix

sample_matrix_2 :: CLMatrix Integer

A sample non-square matrix

sample_matrix_3 :: CLMatrix Integer

Another sample non-square matrix

Ideals and related types

sample_CLReal :: Int -> FPReal

A sample CLReal.

sample_Ideal :: CLIntP -> Ideal

A sample IdealX.

sample_IdealQ :: CLIntP -> IdealQ

A sample IdealQ.

sample_IdealRed :: CLIntP -> IdealRed

A sample IdealRedX.

sample_IdealRedQ :: CLIntP -> IdealRedQ

A sample IdealRedQ.

sample_IdDist :: CLIntP -> IdDist

A sample IdDist.

sample_IdDistQ :: CLIntP -> IdDistQ

A sample IdDistQ.

sample_IdRedDist :: CLIntP -> IdRedDist

A sample IdRedDist.

sample_IdRedDistQ :: CLIntP -> IdRedDistQ

A sample IdRedDistQ.

Testing routines

Smith reduction

test_SNF :: IO ()

Test the Smith Normal Form code.

Class group functions

period_of_ideals :: (IdDist -> IdDist) -> IdDist -> (CLReal, [IdDist])

Classical period finding (just compare the "next" ideal to O and see if it is the same). Takes in the O ideal with appropriate Δ, and returns the circle length (sum δ(I)) and the list of ideals in the first iteration.

show_period_for_bigD :: CLIntP -> String

Show period string for a given Δ.

show_period_for_many_bigDs :: Int -> IO ()

Show the period for the first n valid Δ's.

show_period_for_some_bigD :: CLIntP -> IO ()

Show period string and the list of ideals for a given Δ.

show_bigDs :: Int -> IO ()

Show a list of valid Δ's.

first_few :: IO ()

Explicitly compute first few ideals for some Δ.

op_all_ideals :: (IdDist -> IdDist -> IdDist) -> String -> CLIntP -> IO ()

Perform an operation on all ideal pairs that are generated by Δ.

dot_all_ideals :: CLIntP -> IO ()

The the product of all pairs of ideals for a given Δ.

star_all_ideals :: CLIntP -> IO ()

Take the star product of all pairs of ideals for a given Δ.

test_bounded_while :: (Show int, Integral int) => int -> int -> IO ()

Test the bounded_while functionality.

main :: IO ()

Run classical tests for Class Number algorithm.

test_primes :: IO ()

Test the primes code.