Quipper.Algorithms.QLS.RealFunc

Description

Define various analytic functions for FDouble and QDouble.

Some analytic functions on FDouble.

approx_sin :: FDouble -> FDouble

Approximation of the sine function using Taylor series.

template_approx_sin :: Circ (QDouble -> Circ QDouble)

local_sin :: FDouble -> FDouble

Implementation of the sine function valid on the whole domain of FDouble.

template_local_sin :: Circ (QDouble -> Circ QDouble)

local_cos :: FDouble -> FDouble

Implementation of the cosine function valid on the whole domain of FDouble.

template_local_cos :: Circ (QDouble -> Circ QDouble)

listAngle :: [Double]

template_listAngle :: Circ [QDouble]

listCos :: [Double]

template_listCos :: Circ [QDouble]

listSin :: [Double]

template_listSin :: Circ [QDouble]

list_values :: [(FDouble, FDouble, FDouble)]

template_list_values :: Circ [(QDouble, QDouble, QDouble)]

approx_sqrt :: Int -> FDouble -> FDouble

Auxiliary function for local_sqrt.

template_approx_sqrt :: Circ (Int -> Circ (QDouble -> Circ QDouble))

local_sqrt :: FDouble -> FDouble

Approximation of the square root using iterative means.

template_local_sqrt :: Circ (QDouble -> Circ QDouble)

local_mag :: FDouble -> FDouble -> FDouble

The function magnitude defined for FDouble. Calculate the non-negative magnitude of a complex number.

template_local_mag :: Circ (QDouble -> Circ (QDouble -> Circ QDouble))

rotate :: FDouble -> FDouble -> FDouble -> FDouble -> FDouble -> FDouble -> (FDouble, FDouble)

Apply the matrix

 ( a b )
 ( c d )

to the column vector (x,y).

template_rotate :: Circ (QDouble -> Circ (QDouble -> Circ (QDouble -> Circ (QDouble -> Circ (QDouble -> Circ (QDouble -> Circ (QDouble, QDouble)))))))

approx_atan2_aux :: FDouble -> FDouble -> (FDouble, FDouble, FDouble) -> (FDouble, FDouble, FDouble) -> (FDouble, FDouble, FDouble)

Auxiliary function for approx_atan2.

template_approx_atan2_aux :: Circ (a -> Circ (QDouble -> Circ ((QDouble, QDouble, QDouble) -> Circ ((QDouble, QDouble, QDouble) -> Circ (QDouble, QDouble, QDouble)))))

approx_atan2 :: FDouble -> FDouble -> FDouble

Definition of atan2 using a CORDIC method. Assume (x,y) is in first quadrant and that x > y.

template_approx_atan2 :: Circ (QDouble -> Circ (QDouble -> Circ QDouble))

local_atan2 :: FDouble -> FDouble -> FDouble

Definition of atan2 using a CORDIC method. x and y can be any FDouble.

template_local_atan2 :: Circ (QDouble -> Circ (QDouble -> Circ QDouble))

local_mkPolar :: FDouble -> FDouble -> (FDouble, FDouble)

The function mkPolar defined for FDouble. Form a complex number from polar components of magnitude and phase.

template_local_mkPolar :: Circ (QDouble -> Circ (QDouble -> Circ (QDouble, QDouble)))

class QFloating a where

A type class for quantum floating-point numbers.

Methods

q_sin :: a -> Circ a

Quantum implementation of the sine function.

q_cos :: a -> Circ a

Quantum implementation of the cosine function.

q_atan2 :: QDouble -> QDouble -> Circ QDouble

Quantum implementation of atan2 on QDouble.

q_magnitude :: (QDouble, QDouble) -> Circ QDouble

Quantum implementation of magnitude on QDouble.

q_mkPolar :: QDouble -> QDouble -> Circ (QDouble, QDouble)

Quantum implementation of mkPolar on QDouble.

q_Re :: (QDouble, QDouble) -> Circ QDouble

Quantum implementation of realPart on QDouble: return the real part of a complex number. A quantum complex is a pair of two QDoubles.

q_Im :: (QDouble, QDouble) -> Circ QDouble

Quantum implementation of imagPart on QDouble: return the imaginary part of a copmlex number. A quantum complex is a pair of two QDoubles.

my_test_fdouble :: IO ()

Template subroutines of analytic functions.

template_sin :: Circ (QDouble -> Circ QDouble)

Template version of sin.

template_cos :: Circ (QDouble -> Circ QDouble)

Template version of cos.

template_atan2 :: Circ (QDouble -> Circ (QDouble -> Circ QDouble))

Template version of atan2.

Template subroutines for dealing with quantum complex number encoded as pairs of QDouble.

template_mkPolar :: Circ (QDouble -> Circ (QDouble -> Circ (QDouble, QDouble)))

Template version of mkPolar.

template_symb_colon_symb_plus_ :: Circ (QDouble -> Circ (QDouble -> Circ (QDouble, QDouble)))

Template version of the constructor :+ of Complex.

template_magnitude :: Circ ((QDouble, QDouble) -> Circ QDouble)

Template version of magnitude.

template_realPart :: Circ ((QDouble, QDouble) -> Circ QDouble)

Template version of realPart.

template_imagPart :: Circ ((QDouble, QDouble) -> Circ QDouble)

Template version of imagPart.

Orphan instances

Floating FDouble

Instance details

Methods

pi :: FDouble

exp :: FDouble -> FDouble

log :: FDouble -> FDouble

sqrt :: FDouble -> FDouble

(**) :: FDouble -> FDouble -> FDouble

logBase :: FDouble -> FDouble -> FDouble

sin :: FDouble -> FDouble

cos :: FDouble -> FDouble

tan :: FDouble -> FDouble

asin :: FDouble -> FDouble

acos :: FDouble -> FDouble

atan :: FDouble -> FDouble

sinh :: FDouble -> FDouble

cosh :: FDouble -> FDouble

tanh :: FDouble -> FDouble

asinh :: FDouble -> FDouble

acosh :: FDouble -> FDouble

atanh :: FDouble -> FDouble

log1p :: FDouble -> FDouble

expm1 :: FDouble -> FDouble

log1pexp :: FDouble -> FDouble

log1mexp :: FDouble -> FDouble

RealFloat FDouble

Instance details

Methods

floatRadix :: FDouble -> Integer

floatDigits :: FDouble -> Int

floatRange :: FDouble -> (Int, Int)

decodeFloat :: FDouble -> (Integer, Int)

encodeFloat :: Integer -> Int -> FDouble

exponent :: FDouble -> Int

significand :: FDouble -> FDouble

scaleFloat :: Int -> FDouble -> FDouble

isNaN :: FDouble -> Bool

isInfinite :: FDouble -> Bool

isDenormalized :: FDouble -> Bool

isNegativeZero :: FDouble -> Bool

isIEEE :: FDouble -> Bool

atan2 :: FDouble -> FDouble -> FDouble