Quipper.Algorithms.USV.Simulate

Description

This module contains functions for simulating and debugging the USV algorithm and some of its subroutines.

Documentation

For the coherent arithmetic subroutines defined in Quipper.Algorithms.USV.USV, compare the classical implementation (implemented using Haskell functions) and the quantum one (simulated).

Testing for subroutine h

h_test :: Int -> IO ()

Given an integer n, print the table comparing the output values of the functions h_classical and h_quantum on every possible vector in (ℤ24n)n.