Quipper.Libraries.ClassicalOptim.Circuit

Description

This module provides a simplified representation of classical circuits.

Simplified circuits

type Wire = Int

The type of wires. A wire is determined by an integer ID.

data Gate

The type of gates.

Constructors

NoOp
No operation.

Init Bool Wire
Initialization.

Cnot Wire [(Wire, Bool)]
Multi-controlled not.

wireOfGate :: Gate -> Maybe Wire

Get the wire acted upon by a gate, if any.

ctlsOfGate :: Gate -> Maybe [(Wire, Bool)]

Get the list of controls, if any.

evalCirc :: Map Wire Bool -> [Gate] -> Map Wire Bool

Evaluate a circuit on a given initial state, and return the final state. A state is represented as a map from wires to booleans.

Simplified Circ monad

data CircState

A data structure to represent a "circuit under construction". This holds the data needed for circuit generation.

Constructors

CS

Fields

circuit :: [Gate]

The circuit so far.

freshWire :: Wire

The next fresh wire.

emptyState :: CircState

The empty state.

data Circ a

A simplified Circ monad.

Constructors

Circ (CircState -> (CircState, a))

Low-level access functions

getFresh :: Circ Wire

Retrieve the next fresh wire.

incrementFresh :: Circ ()

Increment the value of the fresh wire.

addGate :: Gate -> Circ ()

Add a new gate to the circuit.

extractCircuit :: Circ a -> [Gate]

Get the circuit out of the monad.

Higher-level access functions

init :: Bool -> Circ Wire

Initialize a new wire.

cnot :: Wire -> [(Wire, Bool)] -> Circ ()

Add a multi-controlled not gate.

Pretty-printing

These functions are only used for testing.

printCircuit :: Circ a -> IO ()

Pretty-print a circuit as a list of gates.

print_quipperStyle :: Gate -> IO ()

Print a gate as Quipper code.