Quipper.Libraries.ClassicalOptim.QuickCheckAlgExp

Quick-checking

twoExp :: Integral a => a -> Int

Compute 2n.

genBoolList :: Integral a => a -> Gen [Bool]

Generate a list of Bool.

test_args :: Args

Arguments for QuickCheck.

test_truth1 :: Int -> IO ()

First test: truth table to expression to truth table is the identity.

genIntList :: [Int] -> Int -> Gen [Int]

Generate a random list of Ints.

genExp :: [Int] -> Gen Exp

Generate a random expression out of the given variables.

test_truth2 :: Int -> IO Result

Second test: expression to truth table to expression is the identity.