Quipper.Libraries.QuipperASCIIParser.Parse

Description

This module contains the code for parsing the ASCII output from Quipper, into a GatePlus. This program is heavily based on, and heavily borrows from the QCLParser.

Documentation

data GatePlus

A type for gates, plus other circuit output, empty lines, and subroutine defs.

Constructors

G Gate [(Wire, Wiretype)]

I [(Wire, Wiretype)]

O [(Wire, Wiretype)]

EmptyLine

CommentLine String

SubroutineName String

SubroutineShape String

Controllable ControllableFlag

Parsing

string_literal :: ReadP String

Parse a string literal.

escaped_char :: ReadP Char

Parse an escaped character, such as 0, n, \, etc.

int :: ReadP Int

Parse a signless integer. We avoid the usual trick (readS_to_P reads), because this introduces backtracking errors.

double :: ReadP Double

Parse a floating point number. We avoid the usual trick (readS_to_P reads), because this introduces backtracking errors.

commalist :: ReadP a -> ReadP [a]

Parse a comma separated list.

control :: ReadP (Signed Wire, Wiretype)

Parse a control structure.

controls :: ReadP ([Signed Wire], [(Wire, Wiretype)])

Parse a list of controls.

wiretype :: ReadP Wiretype

Parse a Wiretype.

wire :: ReadP (Int, Wiretype)

Parse a wire and its type.

wires :: ReadP [(Int, Wiretype)]

Parse a list of input/output wires and types.

none :: ReadP [(Int, Wiretype)]

Parse the string "none", returning an empty list of input/output wires and types.

inversechar :: ReadP Bool

Consume an optional "*". Return True if consumed, and False otherwise.

label' :: ReadP (Int, String)

Consume a label.

labelchar :: ReadP String

Consumer any character other than ')', ']', or ','.

labelbracket :: ReadP String

Consume an index of the form [...].

labels :: ReadP [(Int, String)]

Consume a list of labels.

box_id :: ReadP BoxId

Consume a BoxId followed by a ']' character.

nocontrolflag :: ReadP NoControlFlag

Consume an optional NoControlFlag, returning False if it isn't present.

ascii_line :: ReadP GatePlus

Parse a single line of ASCII output into a Gate. This function needs to be kept in line with Quipper's ascii_render_gate function.

parse_ascii_line :: String -> Maybe GatePlus

The overall parsing function, reading a line of ASCII output, and producing a GatePlus.