Quipper.Utils.Preview

Description

Platform dependent previewing of PDF files. On Windows, we try to run Acrobat Reader. On Mac, we try to run Preview. On Linux, we try to run Acrobat Reader or Xpdf. If this fails, we abort the program with a platform dependent error message.

Documentation

try_several_IO :: [IO a] -> IO a

Sequentially try one or more IO computations, until one of them completes without an IO error. If all of them fail, re-throw the error from the last one.

system_pdf_viewer :: Double -> String -> IO ()

system_pdf_viewer zoom pdffile: Call a system-specific PDF viewer on pdffile file. The zoom argument is out of 100 and may or may not be ignored by the viewer.

windows_pdf_viewer :: Double -> String -> IO ()

Like system_pdf_viewer, but specialized to Windows.

macos_pdf_viewer :: Double -> String -> IO ()

Like system_pdf_viewer, but specialized to Mac OS.

linux_pdf_viewer :: Double -> String -> IO ()

Like system_pdf_viewer, but specialized to Linux.