Quipper.Utils.RandomSource

Description

This module provides a container type for sources of randomness. This makes it possible for a source of randomness (any instance of the RandomGen class) to be stored in a data structure without having to specify its type explicitly.

Documentation

data RandomSource where

A container type to hold a source of randomness. This can hold any instance of the RandomGen class.

Constructors

RandomSource :: forall g. RandomGen g => g -> RandomSource