Quipper.Utils.Sampling

Description

This module provides functions for generating lists of samples from a range of input values. This is primarily useful for generating test cases. Ranges can be specified for types that are members of the Interval class. Each sampling procedure generates a (finite or infinite) list of values from the range. We provide sampling procedures for

generating the range in its entirety (sample_all)

sampling every nth element from a range (sample_step)

generating a random sample from the range (sample_random)

Interval class

class Interval a where

The Interval class contains types for which an interval of values can be specified by giving a lower bound and an upper bound. Intervals are specified as interval min max, for example:

interval (0,0) (1,2) = [(0,0),(0,1),(0,2),(1,0),(1,1),(1,2)].

Methods

interval :: a -> a -> [a]

Takes a range (min,max) and returns a list of all values with lower bound min and upper bound max.

Zero class

class Zero a where

Types in the Zero class have an "origin", i.e., an element that can conveniently serve as the starting point for intervals.

Methods

zero :: a -> a

Inputs any element of the type and outputs the corresponding "zero" element, for example:

zero ([1,2],3,True) = ([0,0],0,False)

Random class

We extend the class Random with tuples and lists.

class Random a

With a source of random number supply in hand, the Random class allows the programmer to extract random values of a variety of types.

Minimal complete definition: randomR and random.

Minimal complete definition

randomR, random

Functions

sample_all :: Interval a => a -> a -> [a]

sample_all min max: returns a list of all elements from the range (min,max). This is actually just a synonym of interval.

sample_step :: (Integral a, Integral b, Interval c) => a -> b -> c -> c -> [c]

sample_step n k min max: returns every nth element from the range (min,max), starting with the kth element.

sample_random :: (Random a, RandomGen g) => g -> a -> a -> [a]

sample_random g min max: returns an infinite list of random samples from the range (min,max), using the random number generator g.

sample_all0 :: (Zero a, Interval a) => a -> [a]

A variant of sample_all that omits the min argument, and uses the zero element of the type instead.

sample_step0 :: (Integral a, Integral b, Zero c, Interval c) => a -> b -> c -> [c]

A variant of sample_step that omits the min argument, and uses the zero element of the type instead.

sample_random0 :: (Random a, Zero a, RandomGen g) => g -> a -> [a]

A variant of sample_random that omits the min argument, and uses the zero element of the type instead.

Orphan instances

Random ()
0-tuples

Instance details

Methods

randomR :: RandomGen g => ((), ()) -> g -> ((), g)

random :: RandomGen g => g -> ((), g)

randomRs :: RandomGen g => ((), ()) -> g -> [()]

randoms :: RandomGen g => g -> [()]

randomRIO :: ((), ()) -> IO ()

randomIO :: IO ()

Random a => Random [a]
Lists

Instance details

Methods

randomR :: RandomGen g => ([a], [a]) -> g -> ([a], g)

random :: RandomGen g => g -> ([a], g)

randomRs :: RandomGen g => ([a], [a]) -> g -> [[a]]

randoms :: RandomGen g => g -> [[a]]

randomRIO :: ([a], [a]) -> IO [a]

randomIO :: IO [a]

(Random a, Random b) => Random (a, b)
Pairs

Instance details

Methods

randomR :: RandomGen g => ((a, b), (a, b)) -> g -> ((a, b), g)

random :: RandomGen g => g -> ((a, b), g)

randomRs :: RandomGen g => ((a, b), (a, b)) -> g -> [(a, b)]

randoms :: RandomGen g => g -> [(a, b)]

randomRIO :: ((a, b), (a, b)) -> IO (a, b)

randomIO :: IO (a, b)

(Random a, Random b, Random c) => Random (a, b, c)
Triples

Instance details

Methods

randomR :: RandomGen g => ((a, b, c), (a, b, c)) -> g -> ((a, b, c), g)

random :: RandomGen g => g -> ((a, b, c), g)

randomRs :: RandomGen g => ((a, b, c), (a, b, c)) -> g -> [(a, b, c)]

randoms :: RandomGen g => g -> [(a, b, c)]

randomRIO :: ((a, b, c), (a, b, c)) -> IO (a, b, c)

randomIO :: IO (a, b, c)

(Random a, Random b, Random c, Random d) => Random (a, b, c, d)
4-Tuples

Instance details

Methods

randomR :: RandomGen g => ((a, b, c, d), (a, b, c, d)) -> g -> ((a, b, c, d), g)

random :: RandomGen g => g -> ((a, b, c, d), g)

randomRs :: RandomGen g => ((a, b, c, d), (a, b, c, d)) -> g -> [(a, b, c, d)]

randoms :: RandomGen g => g -> [(a, b, c, d)]

randomRIO :: ((a, b, c, d), (a, b, c, d)) -> IO (a, b, c, d)

randomIO :: IO (a, b, c, d)

(Random a, Random b, Random c, Random d, Random e) => Random (a, b, c, d, e)
5-Tuples

Instance details

Methods

randomR :: RandomGen g => ((a, b, c, d, e), (a, b, c, d, e)) -> g -> ((a, b, c, d, e), g)

random :: RandomGen g => g -> ((a, b, c, d, e), g)

randomRs :: RandomGen g => ((a, b, c, d, e), (a, b, c, d, e)) -> g -> [(a, b, c, d, e)]

randoms :: RandomGen g => g -> [(a, b, c, d, e)]

randomRIO :: ((a, b, c, d, e), (a, b, c, d, e)) -> IO (a, b, c, d, e)

randomIO :: IO (a, b, c, d, e)

(Random a, Random b, Random c, Random d, Random e, Random f) => Random (a, b, c, d, e, f)
6-Tuples

Instance details

Methods

randomR :: RandomGen g => ((a, b, c, d, e, f), (a, b, c, d, e, f)) -> g -> ((a, b, c, d, e, f), g)

random :: RandomGen g => g -> ((a, b, c, d, e, f), g)

randomRs :: RandomGen g => ((a, b, c, d, e, f), (a, b, c, d, e, f)) -> g -> [(a, b, c, d, e, f)]

randoms :: RandomGen g => g -> [(a, b, c, d, e, f)]

randomRIO :: ((a, b, c, d, e, f), (a, b, c, d, e, f)) -> IO (a, b, c, d, e, f)

randomIO :: IO (a, b, c, d, e, f)

(Random a, Random b, Random c, Random d, Random e, Random f, Random g) => Random (a, b, c, d, e, f, g)
7-Tuples

Instance details

Methods

randomR :: RandomGen g0 => ((a, b, c, d, e, f, g), (a, b, c, d, e, f, g)) -> g0 -> ((a, b, c, d, e, f, g), g0)

random :: RandomGen g0 => g0 -> ((a, b, c, d, e, f, g), g0)

randomRs :: RandomGen g0 => ((a, b, c, d, e, f, g), (a, b, c, d, e, f, g)) -> g0 -> [(a, b, c, d, e, f, g)]

randoms :: RandomGen g0 => g0 -> [(a, b, c, d, e, f, g)]

randomRIO :: ((a, b, c, d, e, f, g), (a, b, c, d, e, f, g)) -> IO (a, b, c, d, e, f, g)

randomIO :: IO (a, b, c, d, e, f, g)