Quipper.Utils.Template.ErrorMsgQ

Description

This module provides a simple monad to encapsulate error messages within the Q monad.

Documentation

type ErrMsg a = Either String a

Shortcut for 'Either String a'.

data ErrMsgQ a

Type for the monad encapsulating error messages.

Constructors

ErrMsgQ (Q (ErrMsg a))

errorMsg :: String -> ErrMsgQ a

Set an error message, to be thrown. Usage:

errorMsg "an error happened"                 

embedQ :: Q a -> ErrMsgQ a

Make a Q computation error-message aware.

extractQ :: String -> ErrMsgQ a -> Q a

Throw the error that has been created, using the given string as a prefix. Usage:

extractQ "name of function: " $ do
  <<commands that may thow an error>>