Quipper.Utils.Template.LiftQ

Description

This module defines the state monad used in Lifting for Template Haskell term manipulation.

Documentation

data LiftState

State of the monad.

Constructors

LiftState

Fields

boundVar :: Map Name Int

How many times each name is bound.

prefix :: Maybe String

The template prefix .

monadName :: Maybe String

The name of the monad.

emptyLiftState :: LiftState

An empty state.

type LiftQState = StateT LiftState ErrMsgQ

Shortcut to StateT LiftState ErrMsgQ.

data LiftQ a

The monad.

Constructors

LiftQ (LiftQState a)

getState :: LiftQ LiftState

Retrieve the state from the monad.

setState :: LiftState -> LiftQ ()

Set the state of the monad.

Various functions to go back and forth between monads.

embedErrMsgQ :: ErrMsgQ a -> LiftQ a

From ErrMsgQ to LiftQ.

embedQ :: Q a -> LiftQ a

From Q to LiftQ.

extractQ :: String -> LiftQ a -> Q a

Get Q out of LiftQ

errorMsg :: String -> LiftQ a

Set an error message.

Working with variable names.

addToBoundVar :: Name -> LiftQ ()

Increase the number of binds of a variable name.

removeFromBoundVar :: Name -> LiftQ ()

Decrease the number of binds of a variable name.

withBoundVar :: Name -> LiftQ a -> LiftQ a

Run a computation with a particular name being bound.

withBoundVars :: [Name] -> LiftQ a -> LiftQ a

Run a computation with a particular list of names being bound.

isBoundVar :: Name -> LiftQ Bool

Say whether a given name is bound.

Other operations on monad state.

setPrefix :: String -> LiftQ ()

Set the template prefix.

getPrefix :: LiftQ String

Get the template prefix.

setMonadName :: String -> LiftQ ()

Set the monad name.

getMonadName :: LiftQ String

Get the monad name.

Functions dealing with variable names.

mkName :: String -> Name

Make a name out of a string.

newName :: String -> LiftQ Name

Make a name out of a string, monadic-style.

sanitizeString :: String -> String

Make any string into a string containing only [0-9a-zA-Z_.]. For example, it replaces any occurrence of "+" with "symb_plus_".

templateString :: String -> LiftQ String

Take a string and make it into a valid Haskell name starting with "template_".

lookForTemplate :: Name -> LiftQ (Maybe Name)

Look for the corresponding "template" name.

makeTemplateName :: Name -> LiftQ Name

Make a the template version of a given name.

Other functions.

prettyPrint :: Ppr a => LiftQ a -> IO ()

Print on the terminal a monadic, printable object.

clauseGetPats :: Clause -> [Pat]

Project patterns out of a clause.

equalNEListElts :: Eq a => [a] -> Bool

Check that the list is a non-empty repetition of the same element.

clausesLengthPats :: [Clause] -> LiftQ Int

Returns the length of the patterns in a list of clauses. Throw an error if the patterns do not have all the same size.