QuipperLib.Qureg

Description

This module provides a data type of quantum registers, as well as associated types of classical and boolean registers.

Quantum registers

type Qureg = Register Qubit

The type of quantum registers. A quantum register is an array of qubits, indexed by natural numbers in the range from 0 to n-1, where n is the length of the register. The syntax a .!(i) is used to access the ith element of the register a.

The main advantage of a register over a list is constant-time access. The main disadvantage is that registers don't allow easy appending, pattern matching, or recursion.

qureg_of_qulist_te :: Qulist -> Qureg

Convert a Qulist to a Qureg. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

qulist_of_qureg_te :: Qureg -> Qulist

Convert a Qureg to a Qulist. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

qureg_length :: Qureg -> Int

Return the length of a Qureg.

qinit_register :: [Bool] -> Circ Qureg

Creates a new quantum register, initialized from a list of booleans. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

qterm_register :: [Bool] -> Qureg -> Circ ()

Terminates a quantum register, and assert that its state is as specified by the list of booleans. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

qmeasure_register :: Qureg -> Circ [Bit]

Measure a quantum register, yielding a list of Bits.

with_ancilla_reg :: Int -> (Qureg -> Circ a) -> Circ a

Temporarily create a quantum register of size n for use as an ancilla. This can be used to introduce an ancilla with a local scope, like this:

 with_ancilla_reg n $ \r -> do {
   <<<code block using ancilla register r>>>
 }

with_ancilla_reg_init :: Boollist -> (Qureg -> Circ a) -> Circ a

Like with_ancilla_reg, except also initialize the register as specified by a bit vector. In this case, the argument n is not required, because it equals the length of the bit vector. When the ancilla is terminated at the end of its scope, it is asserted to be in the same state it was prepared in.

qureg_shape :: Int -> Qureg

Return a piece of shape data to represent an m-qubit quantum register. Please note that the data can only be used as shape; it will be undefined at the leaves.

Bit registers

type Bitreg = Register Bit

The type of Bit registers. The syntax a .!(i) is used to access the ith element of the register a.

bitreg_of_bitlist_te :: Bitlist -> Bitreg

Turn a bit vector into a bit register. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

bitlist_of_bitreg_te :: Bitreg -> Bitlist

Turn a bit register into a bit vector. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

bitreg_length :: Bitreg -> Int

Return the length of a Bitreg.

Boolean registers

type Boolreg = Register Bool

The type of boolean registers.

boolreg_of_boollist_te :: Boollist -> Boolreg

Turn a bool vector into a bool register. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

boollist_of_boolreg_te :: Boolreg -> Boollist

Turn a bool register into a bool vector. The conversion is tail-endian, i.e., r.!(0) holds the tail of the list.

boolreg_length :: Boolreg -> Int

Return the length of a Boolreg.

boolreg_of_int_le :: Integral a => Int -> a -> Boolreg

boolreg_of_int m x: Initialize a bool register directly from an integer x, regarded as a binary string of length m. The conversion is little-endian, i.e., the register holds the least significant digit at index 0.

int_of_boolreg_unsigned_le :: Integral a => Boolreg -> a

int_of_boolreg_unsigned_le m r: Turn a bool register into an integer, regarded as a binary string. The conversion is little-endian, i.e., the register holds the least significant digit at index 0. The integer is unsigned.

General registers

data Register x

A register is an array of elements of some type x, indexed by natural numbers in the range from 0 to n-1, where n is the length of the register.

(.!) :: Register x -> Int -> x

r !.(i): Return the ith element of a register r.